pyproject.toml
[tool.poetry]
name = "designate"
version = "0.0.0"
description = "Retrieve service metadata from Service Registry, render Helm values and trigger Spinnaker pipeline"
authors = ["Dmitry Kovalenko <dmytro.kovalenko@ironsrc.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0.1"
attrs = "^21.2.0"
requests = "^2.25.1"
Jinja2 = "^3.0.1"
deserialize = "^1.8.1"
PyGithub = "^1.55"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
black = "^21.5b2"
isort = "^5.8.0"
flake8 = "^3.9.2"
mypy = "^0.812"
pytest = "^6.2.4"
python-dotenv = "^0.18.0"

[tool.poetry.scripts]
designate = "designate.cli:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
